#297 refactor(shutdown): Put the mocks in their own files
