+++ services/shutdown/src/services/mqtt_test/mock_factory.go
@@ -1,0 +1,12 @@
+package mqtt_test
+
+import (
+	MQTT "github.com/eclipse/paho.mqtt.golang"
+)
+
+type MockFactory struct {
+}
+
+func (factory MockFactory) BuildClient(options *MQTT.ClientOptions) MQTT.Client {
+	return &MockMqttClient{}
+}
